+++ packages/basic/ignores/common.js
@@ -1,5 +1,4 @@
 export default [
-  '**/*.md/*.ts', // disabled for now: not correctly recognized by the `include` block in `tsconfig.eslint.json`
   '**/dist/**',
   '**/dist-ssr/**',
   '**/coverage/**',
@@ -9,7 +8,7 @@ export default [
   '**/tmp/**',
   '*.min.*',
   '*.d.ts',
-  'CHANGELOG.md',
+  'CHANGELOG*',
   'LICENSE*',
   'package-lock.json',
   'pnpm-lock.yaml',