+++ packages/typescript/rules/eslint-comments.js
@@ -1,0 +1,3 @@
+export default {
+  'eslint-comments/disable-enable-pair': ['warn', { allowWholeFile: true }],
+};