+++ packages/typescript/CHANGELOG.md
@@ -1,68 +1,109 @@
 # @williamthorsen/eslint-config-typescript
 
+## 0.9.0
+
+### Minor Changes
+
+- BREAKING CHANGE: Dropped support for ESLint nested config.
+- BREAKING CHANGE: Renamed `@williamthorsen/eslint-config-typescript-flat` to `@williamthorsen/eslint-config-typescript`
+
 ## 0.8.0
 
 ### Minor Changes
 
-- This package is deprecated. Please use v0.9.0 or higher, which uses the ESLint flat config.
-
-### Patch Changes
-
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.8.0
+- This package is deprecated. Please use @williamthorsen/eslint-config-typescript.
 
 ## 0.7.5
 
 ### Patch Changes
 
 - Remove unused import plugin
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.7.5
 
 ## 0.7.4
 
 ### Patch Changes
 
 - Upgrade all dependencies to latest version
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.7.4
 
 ## 0.7.3
 
 ### Patch Changes
 
 - Upgrade all dependencies to latest version
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.7.3
 
 ## 0.7.2
 
 ### Patch Changes
 
 - Upgrade all dependencies to latest version
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.7.2
 
 ## 0.7.1
 
 ### Patch Changes
 
 - Upgrade all dependencies
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.7.1
 
 ## 0.7.0
 
-### Patch Changes
+### Minor Changes
 
 - Upgrade dependencies
-- Updated dependencies
-  - @williamthorsen/eslint-config-basic@0.7.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Replace sort-imports with simple-sort-imports
+
+## 0.6.0
+
+### Minor Changes
+
+- Export common ignore patterns
+
+## 0.5.0
+
+### Minor Changes
+
+- Enforce use of braces for all control statements.
+  Disable no-use-before-define in TypeScript files.
+
+## 0.4.1
+
+### Patch Changes
+
+- Removed erroneous commonIgnores reference from TS rules.
+
+## 0.4.0
+
+### Minor Changes
+
+- Combined JS/TS rules now ignore commonly ignored files
+  Disabled `no-undef` rule in TypeScript files
+
+## 0.3.0
+
+### Minor Changes
+
+- Simplify structure.
+  Make plugins direct dependencies of flat TypeScript config.
+  Remove legacy config.
+
+## 0.2.0
+
+### Minor Changes
+
+- Export a legacy config with same rules as flat config
+
+## 0.1.7
+
+### Patch Changes
+
+- Fix nonpropagating ignores, Node globals
 
 ## 0.1.6
 
 ### Patch Changes
 
-- Upgrade all dependencies to latest version
 - Updated dependencies
   - @williamthorsen/eslint-config-basic@0.1.6